Move device and calendar settings to focused tabs
