+++ frontend/src/components/settings/tabs/system/DebugTab.vue
@@ -28,7 +28,6 @@
         </select>
       </SettingItem>
     </CollapsibleSection>
-
     <CollapsibleSection title="Dashboard Refresh" icon="🔄">
       <SettingItem
         label="Config Polling Interval (seconds)"
@@ -45,24 +44,6 @@
           placeholder="30"
           aria-label="Config polling interval in seconds"
           @change="handlePollIntervalChange"
-        />
-      </SettingItem>
-
-      <SettingItem
-        label="Calendar Refresh Interval (minutes)"
-        help="How often to refresh calendar data (5-120 minutes)"
-        input-id="calendar-refresh-interval"
-      >
-        <input
-          id="calendar-refresh-interval"
-          v-model.number="calendarRefreshInterval"
-          type="number"
-          min="5"
-          max="120"
-          step="1"
-          placeholder="15"
-          aria-label="Calendar refresh interval in minutes"
-          @change="handleCalendarRefreshChange"
         />
       </SettingItem>
     </CollapsibleSection>
@@ -86,9 +67,6 @@ const emit = defineEmits(["update:config"]);
 const consoleLogEnabled = ref(props.config.consoleLogEnabled ?? true);
 const consoleLogLevel = ref(props.config.consoleLogLevel || "info");
 const configPollInterval = ref(props.config.configPollInterval || 30);
-const calendarRefreshInterval = ref(
-  props.config.calendarRefreshInterval || 15,
-);
 
 watch(
   () => props.config,
@@ -96,8 +74,6 @@ watch(
     consoleLogEnabled.value = newConfig.consoleLogEnabled ?? true;
     consoleLogLevel.value = newConfig.consoleLogLevel || "info";
     configPollInterval.value = newConfig.configPollInterval || 30;
-    calendarRefreshInterval.value =
-      newConfig.calendarRefreshInterval || 15;
   },
   { deep: true },
 );
@@ -112,13 +88,6 @@ const handleConsoleLogChange = () => {
 const handlePollIntervalChange = () => {
   emit("update:config", {
     configPollInterval: configPollInterval.value,
-  });
-};
-
-const handleCalendarRefreshChange = () => {
-  const value = Math.max(5, Math.min(120, calendarRefreshInterval.value));
-  emit("update:config", {
-    calendarRefreshInterval: value,
   });
 };
 </script>
