+++ frontend/src/components/settings/tabs/system/PowerTab.vue
@@ -131,53 +131,6 @@
         </div>
       </SettingItem>
     </CollapsibleSection>
-
-    <CollapsibleSection title="Reboot Combo" icon="⌨️">
-      <SettingItem label="First Key" help="First key for reboot combo">
-        <select v-model="rebootComboKey1" @change="handleRebootComboChange">
-          <option value="KEY_1">KEY_1</option>
-          <option value="KEY_2">KEY_2</option>
-          <option value="KEY_3">KEY_3</option>
-          <option value="KEY_4">KEY_4</option>
-          <option value="KEY_5">KEY_5</option>
-          <option value="KEY_6">KEY_6</option>
-          <option value="KEY_7">KEY_7</option>
-        </select>
-      </SettingItem>
-
-      <SettingItem label="Second Key" help="Second key for reboot combo">
-        <select v-model="rebootComboKey2" @change="handleRebootComboChange">
-          <option value="KEY_1">KEY_1</option>
-          <option value="KEY_2">KEY_2</option>
-          <option value="KEY_3">KEY_3</option>
-          <option value="KEY_4">KEY_4</option>
-          <option value="KEY_5">KEY_5</option>
-          <option value="KEY_6">KEY_6</option>
-          <option value="KEY_7">KEY_7</option>
-        </select>
-      </SettingItem>
-
-      <SettingItem
-        label="Combo Duration (milliseconds)"
-        help="How long to hold both keys to trigger reboot (1000-60000 ms)"
-      >
-        <input
-          v-model.number="rebootComboDuration"
-          type="number"
-          min="1000"
-          max="60000"
-          step="1000"
-          @change="handleRebootComboChange"
-        />
-      </SettingItem>
-
-      <SettingItem>
-        <span class="help-text">
-          Hold {{ rebootComboKey1 }} + {{ rebootComboKey2 }} for
-          {{ (rebootComboDuration / 1000).toFixed(1) }} seconds to reboot
-        </span>
-      </SettingItem>
-    </CollapsibleSection>
   </div>
 </template>
 
@@ -219,9 +172,6 @@ const displaySchedule = ref(
 const timezone = ref(props.config.timezone || null);
 const displayTimeoutEnabled = ref(props.config.displayTimeoutEnabled || false);
 const displayTimeout = ref(props.config.displayTimeout || 0);
-const rebootComboKey1 = ref(props.config.rebootComboKey1 || "KEY_1");
-const rebootComboKey2 = ref(props.config.rebootComboKey2 || "KEY_7");
-const rebootComboDuration = ref(props.config.rebootComboDuration || 10000);
 
 const { turnDisplayOn, turnDisplayOff } = useSystem();
 
@@ -234,9 +184,6 @@ watch(
     timezone.value = newConfig.timezone || null;
     displayTimeoutEnabled.value = newConfig.displayTimeoutEnabled || false;
     displayTimeout.value = newConfig.displayTimeout || 0;
-    rebootComboKey1.value = newConfig.rebootComboKey1 || "KEY_1";
-    rebootComboKey2.value = newConfig.rebootComboKey2 || "KEY_7";
-    rebootComboDuration.value = newConfig.rebootComboDuration || 10000;
   },
   { deep: true },
 );
@@ -266,14 +213,6 @@ const handleTimeoutEnabledChange = () => {
 
 const handleTimeoutChange = () => {
   emit("update:config", { displayTimeout: displayTimeout.value });
-};
-
-const handleRebootComboChange = () => {
-  emit("update:config", {
-    rebootComboKey1: rebootComboKey1.value,
-    rebootComboKey2: rebootComboKey2.value,
-    rebootComboDuration: rebootComboDuration.value,
-  });
 };
 
 const handleTurnDisplayOn = async () => {
